+++ CHANGELOG.md
@@ -2,9 +2,13 @@
 
 ---
 
+#### 7.01.2019
+
+* [YouTubeData] Reduced unnecessary service.build() calls
+
 #### 6.01.2019
 
-* [YouTubeData] Fixed a bug in video search  
+* [YouTubeData] Fixed a bug in video search
 * [General] Converted YouTubeData to Kotlin
 
 #### 5.01.2019